+++ src/Commands/owner/debug.js
@@ -1,7 +1,8 @@
 const Command = require('../../Structures/Command');
 
-const { readdirSync, existsSync } = require('fs');
-const { bot: { owner }, emoji: { success, error }, response: { invalidPermissions } } = require('../../../config/config.json');
+const { readdirSync } = require('fs');
+const { bot: { ownerID }, emoji: { success, error }, response: { invalidPermissions } } = require('../../../config/config.json');
+
 
 module.exports = new Command({
 	name: 'debug',
@@ -9,9 +10,18 @@ module.exports = new Command({
 	syntax: 'debug <logFile>',
 	description: 'Sends selected log as a message attachment',
 	async run(message, args, client) {
-		if (message.author.id != owner) return message.channel.send(`${error} ${invalidPermissions}`);
-        if (!args[0]) return message.channel.send(`${success} Logs: ${readdirSync('./logs/').join(', ')}`);
-        if (existsSync(`./logs/${args[0]}`)) message.channel.send({ content: `${success} File: ./logs/${args[0]}`, files: [ `./logs/${args[0]}` ] });
-        else message.channel.send(`${error} That file doesn't exist!`);
+		if (message.author.id != ownerID) return message.channel.send(`${error} ${invalidPermissions}`);
+
+		const files = readdirSync('./logs/');
+
+        if (!args[0]) return message.channel.send(`${success} Logs: **${files.join('**, **')}**`);
+
+		for (const file of files) {
+			if (file.startsWith(args[0])) {
+				message.channel.send({ content: `${success} File: **./logs/${file}**`, files: [ `./logs/${file}` ] });
+				return;
+			}
+		}
+        message.channel.send(`${error} That file doesn't exist!`);
 	}
 });