+++ src/Commands/owner/log.js
@@ -2,7 +2,8 @@ const Command = require('../../Structures/Command');
 
 const { appendFile } = require('fs');
 const { consoleLog } = require('../../Data/Log');
-const { bot: { owner }, emoji: { success, error }, response: { invalidPermissions } } = require('../../../config/config.json');
+const { bot: { ownerID }, emoji: { success, error }, response: { invalidPermissions } } = require('../../../config/config.json');
+
 
 module.exports = new Command({
 	name: 'log',
@@ -10,7 +11,7 @@ module.exports = new Command({
     syntax: 'log [data]',
 	description: 'Manually log into file',
 	async run(message, args, client) {
-        if (message.author.id != owner) return message.channel.send(`${error} ${invalidPermissions}`);
+        if (message.author.id != ownerID) return message.channel.send(`${error} ${invalidPermissions}`);
 		appendFile('./logs/manualLog.txt', `${args.slice(0).join(' ')}\n`, (err) => {
             if (err) {
                 message.channel.send(`${error} Something went wrong :confused: ${err}`);