+++ src/Commands/owner/restart.js
@@ -1,7 +1,7 @@
-const Command = require('../../Structures/Command.js');
+const Command = require('../../Structures/Command');
 
-const { consoleLog } = require('../../Data/Log.js');
-const { bot: { token, owner }, emoji: { success, error, loading }, response: { invalidPermissions } } = require('../../../config/config.json');
+const { consoleLog } = require('../../Data/Log');
+const { bot: { token, ownerID }, emoji: { success, error, loading }, response: { invalidPermissions } } = require('../../../config/config.json');
 
 
 module.exports = new Command({
@@ -9,7 +9,7 @@ module.exports = new Command({
 	aliases: [ 'reboot' ],
 	description: "Restarts the bot's client",
 	async run(message, args, client) {
-		if (message.author.id != owner) return message.channel.send(`${error} ${invalidPermissions}`);
+		if (message.author.id != ownerID) return message.channel.send(`${error} ${invalidPermissions}`);
 
 		consoleLog('[INFO] Restarting...');
 
@@ -17,7 +17,7 @@ module.exports = new Command({
 
 		client.destroy();
 		await client.login(token);
-		
+
 		consoleLog(`[INFO] ${client.user.username} is online and ready on ${client.guilds.cache.size} servers!`);
 		
 		const response = await message.channel.messages.fetch({ limit: 1 });
@@ -26,6 +26,7 @@ module.exports = new Command({
 		}
 		catch (error) {
 			consoleLog(`[WARN] Restart response edit error`, error);
+			message.channel.send(`${success} Restarted!`);
 		}
 	}
 });