+++ src/Commands/owner/setstatus.js
@@ -1,6 +1,6 @@
-const Command = require('../../Structures/Command.js');
+const Command = require('../../Structures/Command');
 
-const { bot: { owner }, status: { status }, emoji: { success, error }, response: { invalidPermissions } } = require('../../../config/config.json');
+const { bot: { ownerID }, status: { status }, emoji: { success, error }, response: { invalidPermissions } } = require('../../../config/config.json');
 
 
 module.exports = new Command({
@@ -9,12 +9,14 @@ module.exports = new Command({
 	syntax: 'setstatus <status>',
 	description: 'Sets the status the bot displays',
 	async run(message, args, client) {
-		if (message.author.id != owner) return message.channel.send(`${error} ${invalidPermissions}`);
-		if ( ![ 'online', 'idle', 'dnd', 'invisible' ].includes(args[0]) ) {
+		if (message.author.id != ownerID) return message.channel.send(`${error} ${invalidPermissions}`);
+		if ( [ 'online', 'idle', 'dnd', 'invisible' ].includes(args[0].toLowerCase()) ) {
+			client.user.setStatus(args[0]);
+        	message.channel.send(`${success} Status set to \`${args[0]}\``);
+		}
+        else {
 			client.user.setStatus(status);
 			return message.channel.send(`${success} Status set to \`${status}\` (default)`);
 		}
-        client.user.setStatus(args[0]);
-        message.channel.send(`${success} Status set to \`${args[0]}\``);
 	}
 });