+++ src/Data/Log.js
@@ -1,38 +1,52 @@
 const { appendFile, writeFile, existsSync, mkdirSync } = require('fs');
-const { logs: { rstLogOnStart } } = require('../../config/config.json');
+const { logs: { resetLogOnStart, logToFile, timeFormat } } = require('../../config/config.json');
 
 
 const initLog = () => {
-    if (!existsSync('./logs')) mkdirSync('./logs');
+    //return new Promise( async (resolve, reject) => {
+        if (!existsSync('./logs')) mkdirSync('./logs');
 
-    if (rstLogOnStart) {
-        writeFile('./logs/sessionLog.txt', '', (err) => {
-            if (err) {
-                return consoleLog(`[WARN] Write file error: `, err);
-            }
-        });
-    }
-    else {
-        for (let i = 0; i < 3; i++) fileLog();
-    }
-    fileLog('[INFO] Process start');
+        if (resetLogOnStart) {
+            writeFile('./logs/sessionLog.txt', '', (err) => {
+                if (err) {
+                    return consoleLog(`[WARN] Write file error: `, err);
+                }
+            });
+        }
+        else if (logToFile) {
+            for (let i = 0; i < 3; i++) fileLog();
+        }
+        fileLog('[INFO] Process start');
+    //    resolve();
+    //});
 }
 
-const fileLog = (message = '', data) => {
+const fileLog = async (message = '', ...data) => {
+    if (!logToFile) { //only log to console then return so log file isn't written to
+        if (data.length > 0) console.log(`${time()} ${message}: `, ...data);
+        else console.log(`${time()} ${message}`);
+        return;
+    }
+
     if (message != '') message = `${time()} ${message}`;
-    appendFile('./logs/sessionLog.txt', `${message}${data ? ': ' : ''}\n`, (err) => {
+    await appendFile('./logs/sessionLog.txt', `${message}${data.length > 0 ? ': ' : ''}\n`, (err) => {
+        // despite VSCode says await has no effect on this function, it does cause it to write in correct order
         if (err) {
             return console.log(`[WARN] Append file error: `, err);
         }
     });
-    if (data) {
+    if (data.length > 0) {
+        let jsonData = '';
         try {
-            data = JSON.stringify(data);
+            data.forEach((element) => {
+                jsonData = jsonData.concat( JSON.stringify(element, null, 4) );
+                jsonData = jsonData.concat('\n');
+            });
         }
         catch(err) {
-            data = `Can\'t stringify data: ${err.message}`;
+            jsonData = `Can\'t stringify data: ${err.message}`;
         }
-        appendFile('./logs/sessionLog.txt', data, (err) => {
+        appendFile('./logs/sessionLog.txt', jsonData, (err) => {
             if (err) {
                 return console.log(`[WARN] Append file error: `, err);
             }
@@ -45,14 +59,15 @@ const fileLog = (message = '', data) => {
     }
 }
 
-const consoleLog = (message, data) => {
-    fileLog(message, data);
-    if (data) console.log(`${time()} ${message}: `, data);
+const consoleLog = (message, ...data) => {
+    fileLog(message, ...data);
+    if (!logToFile) return; // only handle console.log in filelog
+    if (data.length > 0) console.log(`${time()} ${message}: `, ...data);
     else console.log(`${time()} ${message}`);
 }
 
 const time = () => {
-    return `[${new Date().toLocaleString('cs-CZ', { hour: 'numeric', minute: 'numeric', second: 'numeric' })}]`;
+    return `[${new Date().toLocaleString(timeFormat, { hour: 'numeric', minute: 'numeric', second: 'numeric' })}]`;
 }
 
 module.exports = { initLog, fileLog, consoleLog }