+++ src/Structures/Client.js
@@ -28,10 +28,10 @@ class Client extends Discord.Client {
 	}
 
     start() {
-		consoleLog(`This application comes from a GitHub project ${homepage.substring(19, homepage.length - 7)} (${homepage}).\nThe use is possible for free while keeping the credits.\nMade by EnhancedMind\nVersion ${version}\n`);
-
 		initLog();
 		checkSettingsFiles();
+
+		consoleLog(`\nThis application comes from a GitHub project ${homepage.substring(19, homepage.length - 7)} (${homepage}).\nThe use is possible for free while keeping the credits.\nMade by EnhancedMind\nVersion ${version}\n`);
 
 		readdirSync('./src/Commands')
 			.forEach(dirs => {