+++ src/main.js
@@ -1,10 +1,8 @@
 const { terminateOnUncaughtException } = require('../config/config.json');
-const { fileLog } = require('./Data/Log');
 
 if (!terminateOnUncaughtException) {
     process.on('uncaughtException', (err) => {
-        console.log('Uncaught Exception: ', err);
-        fileLog('Uncaught Exception: ', err);
+        console.error('Uncaught Exception: ', err.stack);
     });
 }